testing employee account

+++ website/views.py
@@ -6,13 +6,14 @@ from .models import Attendance
 from django.urls import reverse
 from datetime import datetime
 from decimal import Decimal, ROUND_HALF_UP
-from django.utils import timezone
+# from django.utils import timezone
+# from django.contrib.auth.hashers import check_password
+from .models import Employee
 # Create your views here.
 
 # landing page
 def home(request):
     return render(request, 'home.html',)
-
 
 # login page (admin)
 def login_user(request):
@@ -35,10 +36,9 @@ def welcome_view(request, username):
         return render(request, 'welcome.html', {'username': username})
     else:
         logout(request)
+        messages.success(request, "You must be logged it to access the page")
         return redirect('home')  
     
-
-
 # logout
 def logout_user(request):
     logout(request)
@@ -63,22 +63,29 @@ def register_user(request):
         return render(request, 'register.html',{'form':form})
     return render(request, 'register.html',{'form':form})
 
+
+
 # attendance_list admin
-def attendance_list(request):
-    if request.user.is_authenticated:
+def attendance_list(request, username):
+    if request.user.is_authenticated and request.user.username == username:
         attendances = Attendance.objects.all()
         for attendance in attendances:
             computed_salary = calculate_computed_salary(attendance.time_in, attendance.time_out, attendance.salary)
             attendance.computed_salary = computed_salary
-        return render(request, 'attendanceList.html',{'attendances': attendances})
+        return render(request, 'attendanceList.html',{'attendances': attendances, 'username': username})
     else:
-        messages.success(request, "You Must be logged in to add record")
+        messages.success(request, "You Must be logged in to view the attendace list")
         return redirect('home')
 
+
+
+
+
+
 # create time in (for employees) (add autofill) (add security)  
-def time_In(request):
+def time_In(request, username):
     form = AttendanceForm(request.POST or None)
-    if request.user.is_authenticated:
+    if request.user.is_authenticated and request.user.username == username:
             if request.method == "POST":
                 if form.is_valid():
                     add_record = form.save()
@@ -86,17 +93,16 @@ def time_In(request):
                     return redirect('home')
             return render(request, 'timeIn.html',{'form':form})
     else:
-        messages.success(request, "You Must be logged in to add record")
-        return redirect('home')
+        return redirect('attendanceList', username=username)
     
 
 
  # admin attendance individual record
-def attendance_record(request, pk):
-    if request.user.is_authenticated:
+def attendance_record(request, pk, username):
+    if request.user.is_authenticated and request.user.username == username:
         attendance_record = Attendance.objects.get(id=pk)
         computed_salary = calculate_computed_salary(attendance_record.time_in, attendance_record.time_out, attendance_record.salary)
-        return render(request, 'attendanceRecord.html', {'attendance_record': attendance_record, 'computed_salary': computed_salary})
+        return render(request, 'attendanceRecord.html', {'attendance_record': attendance_record, 'computed_salary': computed_salary, 'username': username})
     else:
         messages.success(request, "You must be logged in to view that page")
         return redirect('home')
@@ -104,20 +110,20 @@ def attendance_record(request, pk):
 
 
     # deletion of record
-def delete_record(request, pk):
-    if request.user.is_authenticated:
+def delete_record(request, pk,username):
+    if request.user.is_authenticated and request.user.username == username:
         delete_it = Attendance.objects.get(id=pk)
         delete_it.delete()
         messages.success(request, "Record Deleted Successfully")
-        return redirect('attendanceList')
+        return redirect('attendanceList',  username=username)
     else:
         messages.success(request, "You Must be logged in to delete this record")
-        return redirect('attendanceList')
+        return redirect('attendanceList', username=username)
     
 
     # time out
-def time_Out(request, pk):
-    if request.user.is_authenticated:
+def time_Out(request, pk, username):
+    if request.user.is_authenticated and request.user.username == username:
         current_record = Attendance.objects.get(id=pk)
         truncated_time_in = truncate_to_minutes(current_record.time_in)
         truncated_time_out = truncate_to_minutes(current_record.time_out)
@@ -134,15 +140,15 @@ def time_Out(request, pk):
             if form.is_valid():
                 form.save()
                 messages.success(request, "Record Has Been Updated!")
-                return redirect('attendanceList')
+                return redirect('attendanceList', {'username': username})
             
             return render(request, 'timeOut.html', {'form': form, 'computed_salary': computed_salary})
         else:
             messages.warning(request, "Record already timed out.")
-            return redirect('attendanceList')
+            return redirect('attendanceList', {'username': username})
     else:
         messages.success(request, "You Must Be Logged In...")
-        return redirect('attendanceList')
+        return redirect('home')
 
 
 def truncate_to_minutes(time):
@@ -164,3 +170,25 @@ def calculate_computed_salary_timeout(time_in, salary):
     computed_salary = (salary_decimal / Decimal(9)) * Decimal(hours_difference)
     computed_salary = computed_salary.quantize(Decimal('0.01'), rounding=ROUND_HALF_UP)
     return computed_salary
+
+
+
+# Emloyeee functions
+
+
+def login_Employee(request):
+    if request.method == 'POST':
+        username = request.POST['username']
+        password = request.POST['password']
+        employee = authenticate(request, username=username, password=password)
+        if employee is not None:
+            login(request, employee)
+            messages.success(request, "You have been logged in!")
+            return redirect(reverse('welcome', kwargs={'username': username}))
+        else:
+            messages.error(request, "Invalid credentials")
+            return redirect('employeeLogin')
+    else:
+        return render(request, 'employeeLogin.html')
+    
+
